refactored GPSService into ServiceCommand
