+++ app/src/main/java/com/njackson/application/modules/AndroidModule.java
@@ -12,7 +12,7 @@ import com.njackson.activityrecognition.ActivityRecognitionService;
 import com.njackson.application.PebbleBikeApplication;
 import com.njackson.activities.SettingsActivity;
 import com.njackson.fragments.StartButtonFragment;
-import com.njackson.gps.GPSService;
+import com.njackson.gps.GPSServiceCommand;
 import com.njackson.oruxmaps.IOruxMaps;
 import com.njackson.oruxmaps.OruxMaps;
 import com.njackson.oruxmaps.OruxMapsService;
@@ -38,7 +38,7 @@ import static android.content.Context.SENSOR_SERVICE;
 /**
  * Created by server on 30/03/2014.
  */
-@Module(library = true,complete=false,injects = {GPSService.class, ActivityRecognitionService.class, MainActivity.class, SettingsActivity.class, StartButtonFragment.class,PebbleService.class, OruxMapsService.class, PebbleDataReceiver.class})
+@Module(library = true,complete=false,injects = {GPSServiceCommand.class, ActivityRecognitionService.class, MainActivity.class, SettingsActivity.class, StartButtonFragment.class,PebbleService.class, OruxMapsService.class, PebbleDataReceiver.class})
 public class AndroidModule {
     private final PebbleBikeApplication application;
