+++ app/src/main/java/com/njackson/events/GPSServiceCommand/GPSStatus.java
@@ -1,4 +1,4 @@
-package com.njackson.events.status;
+package com.njackson.events.GPSServiceCommand;
 
 /**
  * Created by server on 21/03/2014.
