+++ app/src/main/java/com/njackson/pebble/PebbleService.java
@@ -12,8 +12,8 @@ import com.njackson.adapters.NewLocationToPebbleDictionary;
 import com.njackson.application.PebbleBikeApplication;
 import com.njackson.events.PebbleService.NewMessage;
 import com.njackson.events.status.PebbleStatus;
-import com.njackson.events.status.GPSStatus;
-import com.njackson.events.GPSService.NewLocation;
+import com.njackson.events.GPSServiceCommand.GPSStatus;
+import com.njackson.events.GPSServiceCommand.NewLocation;
 import com.njackson.events.LiveService.LiveMessage;
 import com.njackson.pebble.canvas.CanvasPlugin;
 import com.njackson.pebble.canvas.GPSData;
